Move send channel sample to "Producing" project
