+++ DotPulsar.sln
@@ -28,8 +28,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Compression", "benchmarks\C
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Benchmarks", "Benchmarks", "{2C57AF4B-0D23-42D7-86FE-80277FD52875}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SendChannel", "samples\SendChannel\SendChannel.csproj", "{366ACA6A-7DA2-4E51-AC27-7B570DFFE5D2}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Extensions", "samples\Extensions\Extensions.csproj", "{F211960B-F2C6-4878-B1AD-72D63CDB3B7A}"
 EndProject
 Global
@@ -62,10 +60,6 @@ Global
 		{040F8253-074D-4977-BDB1-0D9798B52CE2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{040F8253-074D-4977-BDB1-0D9798B52CE2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{040F8253-074D-4977-BDB1-0D9798B52CE2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{366ACA6A-7DA2-4E51-AC27-7B570DFFE5D2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{366ACA6A-7DA2-4E51-AC27-7B570DFFE5D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{366ACA6A-7DA2-4E51-AC27-7B570DFFE5D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{366ACA6A-7DA2-4E51-AC27-7B570DFFE5D2}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F211960B-F2C6-4878-B1AD-72D63CDB3B7A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F211960B-F2C6-4878-B1AD-72D63CDB3B7A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F211960B-F2C6-4878-B1AD-72D63CDB3B7A}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -80,7 +74,6 @@ Global
 		{14934BED-A222-47B2-A58A-CFC4AAB89B49} = {E7106D0F-B255-4631-9FB8-734FC5748FA9}
 		{6D44683B-865C-4D15-9F0A-1A8441354589} = {E7106D0F-B255-4631-9FB8-734FC5748FA9}
 		{040F8253-074D-4977-BDB1-0D9798B52CE2} = {2C57AF4B-0D23-42D7-86FE-80277FD52875}
-		{366ACA6A-7DA2-4E51-AC27-7B570DFFE5D2} = {E7106D0F-B255-4631-9FB8-734FC5748FA9}
 		{F211960B-F2C6-4878-B1AD-72D63CDB3B7A} = {E7106D0F-B255-4631-9FB8-734FC5748FA9}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
